Added code to prevent the ssr from trying to access localStorage

+++ src/pages/page-creation.jsx
@@ -40,9 +40,11 @@ const tabIndexMap = {
   "Growth Report": 2,
 };
 let defaultTab = null;
-if (localStorage.getItem("defaultTabIndex") != null) {
-  defaultTab = parseInt(localStorage.getItem("defaultTabIndex"));
-  localStorage.clear();
+if (typeof window !== "undefined") {
+  if (localStorage.getItem("defaultTabIndex") != null) {
+    defaultTab = parseInt(localStorage.getItem("defaultTabIndex"));
+    localStorage.clear();
+  }
 }
 const PageCreation = () => {
   const { isOpen, onOpen, onClose } = useDisclosure();
